(#2) refactor: introduce `SingleThreadEventLoop` abstraction and simplify `TaskEventLoop` implementation by delegating shared behavior
